réparation de quelques exercices

+++ libft.h
@@ -6,7 +6,7 @@
 /*   By: dan <dan@student.42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/01 20:44:41 by dan               #+#    #+#             */
-/*   Updated: 2022/10/03 18:08:29 by dan              ###   ########.fr       */
+/*   Updated: 2022/10/03 18:44:39 by dan              ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,7 +26,7 @@ void	*ft_memcpy(void *dest, const void *src, size_t n);
 void	*ft_memmove(void *dest, const void *src, size_t n);
 int		ft_strlcat(char *dest, char *src, size_t size);
 int		ft_strlcpy(char *dest, char *src, size_t size);
-void	*ft_memset(void *str, int c, int n);
+void	*ft_memset(void *str, int c, size_t n);
 size_t	ft_strlen(char *str);
 int		ft_toupper(int c);
 int		ft_tolower(int c);
